Fixed overwriting csv
I was getting the csv with only the ('Title', 'Directors', 'Year') row. In line 20 the code was overwriting the file generated by the parse function. I've changed to order, so the script first create the csv and then append the rows.

And make a few tweaks in the last lines to accept only input that start with 's' or 'n'